api/requirements.txt
# MONAI and Medical Imaging Dependencies
monai>=1.3.0
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pydicom>=2.4.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Additional ML libraries
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Data processing
pandas>=2.0.0

# Web framework dependencies
flask>=2.3.0
gunicorn>=21.2.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0

# Authentication
PyJWT>=2.8.0
cryptography>=41.0.0
